Use ruamel.yaml instead of pyyaml for 40x performance boost
pyyaml is not distributed with the C libyaml, so by default uses a
pure python yaml parser which is very slow.

I tried to get the lambda to package with pyyaml+libyaml, but couldn't
make it work. Switching to ruamel.yaml was easy, and it is available
as manylinux binary wheels.

This switch drops the approximate cost from $62/1M Invocations to
$1.46/1M Invocation. (Each invocation processes 10 conversions).