Fix security bug with raijin script execution and simplify scripts dir
It was previously been possible to execute any file on raijin,
include scripts writable by arbitrary users, by passing ../../ etc
as part of the command name. This was still protected by using safe
SSH keys, but circumvented the intended white list of available
commands.

Most of the scripts we have are a single script, and I find it awkward
that they must be contained within a subdirectory and all the the same
name. This PR allows following the current convention, or having a
single file be the runnable script.