Perform module load step inside subshell...
Environment variables are seemingly lost when launching the subshell,
this should not be, but I don't know why.

Also append to the log file instead of replace, so that we keep
all the log messages.

If we're going to log like this, we should use tee so that it is stored
in the filesystem as well as in CloudWatch Logs.